% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinycodetfakepkg2.R
\name{tinycodetfakepkg2}
\alias{tinycodetfakepkg2}
\alias{fun_overwritten}
\alias{fun21}
\alias{fun22}
\alias{\%opover\%}
\alias{\%op21\%}
\alias{\%op22\%}
\title{tinycodetfakepkg2ellaneous functions to help your coding etiquette}
\usage{
fun_overwritten()

fun21()

fun22()

X \%opover\% A

X \%op21\% A

X \%op22\% A
}
\arguments{
\item{X}{whatever}

\item{A}{whatever}
}
\description{
Bla
}
